Refactor dictation workflow seam and clarify AppModel ownership
